requirement.txt
torch
torchvision
opencv-python
Pillow
requests
tk
